requirements.txt
fastapi
uvicorn
kafka-python
fpdf
pandas
matplotlib
seaborn
openpyxl
telegram